+++ apps/meteor/client/views/navigation/sidepanel/SidePanelRouter.tsx
@@ -1,7 +1,5 @@
-import { usePermission, useSetting } from '@rocket.chat/ui-contexts';
-
 import SidePanelAll from './tabs/SidePanelAll';
-import { ALL_GROUPS, useRoomsListContext, useSidePanelFilter, useSwitchSidePanelTab } from '../contexts/RoomsNavigationContext';
+import { ALL_GROUPS, useRoomsListContext, useSidePanelFilter } from '../contexts/RoomsNavigationContext';
 import SidePanelDiscussions from './tabs/SidePanelDiscussions';
 import SidePanelFavorites from './tabs/SidePanelFavorites';
 import SidePanelInProgress from './tabs/SidePanelInProgress';
@@ -9,22 +7,10 @@ import SidePanelMentions from './tabs/SidePanelMentions';
 import SidePanelQueue from './tabs/SidePanelQueue';
 import SidePanelRooms from './tabs/SidePanelRooms';
 import SidePanelOnHold from './tabs/SidepanelOnHold';
-import { useHasLicenseModule } from '../../../hooks/useHasLicenseModule';
 
 const SidePanelRouter = () => {
 	const [currentTab] = useSidePanelFilter();
 	const { parentRid } = useRoomsListContext();
-	const switchSidePanelTab = useSwitchSidePanelTab();
-
-	const isDiscussionEnabled = useSetting('Discussion_enabled');
-	const hasEEModule = useHasLicenseModule('livechat-enterprise');
-	const canViewOmnichannelQueue = usePermission('view-livechat-queue');
-	const isQueueEnabled = useSetting('Livechat_waiting_queue');
-
-	const resetSidePanelTab = () => {
-		switchSidePanelTab(ALL_GROUPS.ALL);
-		return null;
-	};
 
 	switch (currentTab) {
 		case ALL_GROUPS.ALL:
@@ -34,10 +20,7 @@ const SidePanelRouter = () => {
 		case ALL_GROUPS.FAVORITES:
 			return <SidePanelFavorites />;
 		case ALL_GROUPS.DISCUSSIONS:
-			if (isDiscussionEnabled) {
-				return <SidePanelDiscussions />;
-			}
-			return resetSidePanelTab();
+			return <SidePanelDiscussions />;
 		case ALL_GROUPS.TEAMS:
 		case ALL_GROUPS.CHANNELS:
 		case ALL_GROUPS.DIRECT_MESSAGES:
@@ -45,17 +28,11 @@ const SidePanelRouter = () => {
 		case ALL_GROUPS.IN_PROGRESS:
 			return <SidePanelInProgress />;
 		case ALL_GROUPS.ON_HOLD:
-			if (hasEEModule) {
-				return <SidePanelOnHold />;
-			}
-			return resetSidePanelTab();
+			return <SidePanelOnHold />;
 		case ALL_GROUPS.QUEUE:
-			if (hasEEModule && canViewOmnichannelQueue && isQueueEnabled) {
-				return <SidePanelQueue />;
-			}
-			return resetSidePanelTab();
+			return <SidePanelQueue />;
 		default:
-			return resetSidePanelTab();
+			return null;
 	}
 };
