+++ apps/meteor/client/views/navigation/sidepanel/tabs/SidePanelDiscussions.tsx
@@ -1,9 +1,10 @@
-import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
+import { useSetting, type SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
 import { useTranslation } from 'react-i18next';
 
 import {
 	SIDE_PANEL_GROUPS,
 	sidePanelFiltersConfig,
+	useRedirectToDefaultTab,
 	useSidePanelRoomsListTab,
 	useUnreadOnlyToggle,
 } from '../../contexts/RoomsNavigationContext';
@@ -13,6 +14,12 @@ const SidePanelDiscussions = () => {
 	const { t } = useTranslation();
 	const rooms = useSidePanelRoomsListTab(SIDE_PANEL_GROUPS.DISCUSSIONS);
 	const [unreadOnly, toggleUnreadOnly] = useUnreadOnlyToggle();
+	const isDiscussionEnabled = useSetting('Discussion_enabled');
+	useRedirectToDefaultTab(!isDiscussionEnabled);
+
+	if (!isDiscussionEnabled) {
+		return null;
+	}
 
 	return (
 		<SidePanel
