+++ apps/meteor/client/views/navigation/sidepanel/tabs/SidePanelQueue.tsx
@@ -1,9 +1,11 @@
-import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
+import { usePermission, useSetting, type SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
 import { useTranslation } from 'react-i18next';
 
+import { useHasLicenseModule } from '../../../../hooks/useHasLicenseModule';
 import {
 	SIDE_PANEL_GROUPS,
 	sidePanelFiltersConfig,
+	useRedirectToDefaultTab,
 	useSidePanelRoomsListTab,
 	useUnreadOnlyToggle,
 } from '../../contexts/RoomsNavigationContext';
@@ -11,8 +13,19 @@ import SidePanel from '../SidePanel';
 
 const SidePanelQueue = () => {
 	const { t } = useTranslation();
+
+	const hasEEModule = useHasLicenseModule('livechat-enterprise');
+	const canViewOmnichannelQueue = usePermission('view-livechat-queue');
+	const isQueueEnabled = useSetting('Livechat_waiting_queue');
+
 	const rooms = useSidePanelRoomsListTab(SIDE_PANEL_GROUPS.QUEUE);
 	const [unreadOnly, toggleUnreadOnly] = useUnreadOnlyToggle();
+	const shouldDisplayQueue = hasEEModule && canViewOmnichannelQueue && isQueueEnabled;
+	useRedirectToDefaultTab(!shouldDisplayQueue);
+
+	if (!shouldDisplayQueue) {
+		return null;
+	}
 
 	return (
 		<SidePanel
