+++ apps/meteor/client/views/navigation/sidepanel/tabs/SidepanelOnHold.tsx
@@ -1,9 +1,11 @@
 import type { SubscriptionWithRoom } from '@rocket.chat/ui-contexts';
 import { useTranslation } from 'react-i18next';
 
+import { useHasLicenseModule } from '../../../../hooks/useHasLicenseModule';
 import {
 	SIDE_PANEL_GROUPS,
 	sidePanelFiltersConfig,
+	useRedirectToDefaultTab,
 	useSidePanelRoomsListTab,
 	useUnreadOnlyToggle,
 } from '../../contexts/RoomsNavigationContext';
@@ -13,6 +15,13 @@ const SidePanelOnHold = () => {
 	const { t } = useTranslation();
 	const rooms = useSidePanelRoomsListTab(SIDE_PANEL_GROUPS.ON_HOLD);
 	const [unreadOnly, toggleUnreadOnly] = useUnreadOnlyToggle();
+
+	const hasEEModule = useHasLicenseModule('livechat-enterprise');
+	useRedirectToDefaultTab(!hasEEModule);
+
+	if (!hasEEModule) {
+		return null;
+	}
 
 	return (
 		<SidePanel
